refactor: modularize ralph-cli and ralph-core for improved code organization
Extract loop runner and display logic from main.rs into dedicated modules.
Reorganize event_loop.rs into a module directory with separate files for
loop state and tests.
