+++ .agent/tasks.jsonl
@@ -16,3 +16,9 @@
 {"id":"task-1769097899-7306","title":"Add task breakdown guidance to core_prompt()","status":"closed","priority":1,"blocked_by":[],"created":"2026-01-22T16:04:59.488201+00:00","closed":"2026-01-22T16:07:00.660544+00:00"}
 {"id":"task-1769097902-2a3c","title":"Add state management guidance (memories vs context files)","status":"closed","priority":1,"blocked_by":[],"created":"2026-01-22T16:05:02.928319+00:00","closed":"2026-01-22T16:07:05.782544+00:00"}
 {"id":"task-1769097909-2b0e","title":"List .agent/*.md files in prompt for context discovery","status":"closed","priority":1,"blocked_by":[],"created":"2026-01-22T16:05:09.469777+00:00","closed":"2026-01-22T16:07:09.251742+00:00"}
+{"id":"task-1769098990-4da4","title":"Refactor main.rs: Extract run_loop_impl to loop_runner.rs","status":"closed","priority":2,"blocked_by":[],"created":"2026-01-22T16:23:10.347563+00:00","closed":"2026-01-22T16:39:21.298296+00:00"}
+{"id":"task-1769098996-f568","title":"Refactor main.rs: Extract display functions to display.rs","status":"closed","priority":2,"blocked_by":[],"created":"2026-01-22T16:23:16.783726+00:00","closed":"2026-01-22T16:39:21.303152+00:00"}
+{"id":"task-1769098998-46fb","title":"Refactor event_loop.rs: Move tests to event_loop/tests.rs","status":"closed","priority":2,"blocked_by":[],"created":"2026-01-22T16:23:18.214782+00:00","closed":"2026-01-22T16:39:26.334244+00:00"}
+{"id":"task-1769098999-3ebf","title":"Refactor event_loop.rs: Extract LoopState to loop_state.rs","status":"closed","priority":2,"blocked_by":[],"created":"2026-01-22T16:23:19.147141+00:00","closed":"2026-01-22T16:39:26.337904+00:00"}
+{"id":"task-1769099000-7e0a","title":"Refactor state.rs: Extract SearchState to search.rs","status":"closed","priority":3,"blocked_by":[],"created":"2026-01-22T16:23:20.163345+00:00","closed":"2026-01-22T16:39:30.766592+00:00"}
+{"id":"task-1769099000-03c7","title":"Refactor state.rs: Extract IterationBuffer to iteration_buffer.rs","status":"closed","priority":3,"blocked_by":[],"created":"2026-01-22T16:23:20.984011+00:00","closed":"2026-01-22T16:39:30.770311+00:00"}
