+++ crates/ralph-cli/src/loop_runner.rs
@@ -1,0 +1,1012 @@
+//! Core orchestration loop implementation.
+//!
+//! This module contains the main `run_loop_impl` function that executes
+//! the Ralph orchestration loop, along with supporting types and helper
+//! functions for PTY execution and termination handling.
+
+use anyhow::{Context, Result};
+use ralph_adapters::{
+    CliBackend, CliExecutor, ConsoleStreamHandler, OutputFormat as BackendOutputFormat,
+    PrettyStreamHandler, PtyConfig, PtyExecutor, QuietStreamHandler, TuiStreamHandler,
+};
+use ralph_core::{
+    EventLogger, EventLoop, EventParser, EventRecord, RalphConfig, Record, SessionRecorder,
+    SummaryWriter, TerminationReason,
+};
+use ralph_proto::{Event, HatId};
+use ralph_tui::Tui;
+use std::fs::{self, File};
+use std::io::{BufWriter, IsTerminal, stdout};
+use std::path::PathBuf;
+use std::process::Command;
+use std::sync::Arc;
+use std::time::Duration;
+use tracing::{debug, error, info, warn};
+
+use crate::display::{build_tui_hat_map, print_iteration_separator, print_termination};
+use crate::process_management;
+use crate::{ColorMode, Verbosity};
+
+/// Outcome of executing a prompt via PTY or CLI executor.
+pub(crate) struct ExecutionOutcome {
+    pub output: String,
+    pub success: bool,
+    pub termination: Option<TerminationReason>,
+}
+
+/// Core loop implementation supporting both fresh start and continue modes.
+///
+/// `resume`: If true, publishes `task.resume` instead of `task.start`,
+/// signaling the planner to read existing scratchpad rather than doing fresh gap analysis.
+///
+/// `record_session`: If provided, records all events to the specified JSONL file for replay testing.
+pub async fn run_loop_impl(
+    config: RalphConfig,
+    color_mode: ColorMode,
+    resume: bool,
+    enable_tui: bool,
+    verbosity: Verbosity,
+    record_session: Option<PathBuf>,
+) -> Result<TerminationReason> {
+    // Set up process group leadership per spec
+    // "The orchestrator must run as a process group leader"
+    process_management::setup_process_group();
+
+    let use_colors = color_mode.should_use_colors();
+
+    // Determine effective execution mode (with fallback logic)
+    // Per spec: Claude backend requires PTY mode to avoid hangs
+    // TUI mode is observation-only - uses streaming mode, not interactive
+    let interactive_requested = config.cli.default_mode == "interactive" && !enable_tui;
+    let user_interactive = if interactive_requested {
+        if stdout().is_terminal() {
+            true
+        } else {
+            warn!("Interactive mode requested but stdout is not a TTY, falling back to autonomous");
+            false
+        }
+    } else {
+        false
+    };
+    // Always use PTY for real-time streaming output (vs buffered CliExecutor)
+    let use_pty = true;
+
+    // Set up signal handling for immediate termination
+    // Per spec:
+    // - SIGINT (Ctrl+C): Immediately terminate child process (SIGTERM -> 5s grace -> SIGKILL), exit with code 130
+    // - SIGTERM: Same as SIGINT
+    // - SIGHUP: Same as SIGINT
+    //
+    // Use watch channel for interrupt notification so we can race execution vs interrupt
+    let (interrupt_tx, interrupt_rx) = tokio::sync::watch::channel(false);
+
+    // Spawn task to listen for SIGINT (Ctrl+C)
+    let interrupt_tx_sigint = interrupt_tx.clone();
+    tokio::spawn(async move {
+        if tokio::signal::ctrl_c().await.is_ok() {
+            debug!("Interrupt received (SIGINT), terminating immediately...");
+            let _ = interrupt_tx_sigint.send(true);
+        }
+    });
+
+    // Spawn task to listen for SIGTERM (Unix only)
+    #[cfg(unix)]
+    {
+        let interrupt_tx_sigterm = interrupt_tx.clone();
+        tokio::spawn(async move {
+            let mut sigterm =
+                tokio::signal::unix::signal(tokio::signal::unix::SignalKind::terminate())
+                    .expect("Failed to register SIGTERM handler");
+            sigterm.recv().await;
+            debug!("SIGTERM received, terminating immediately...");
+            let _ = interrupt_tx_sigterm.send(true);
+        });
+    }
+
+    // Spawn task to listen for SIGHUP (Unix only)
+    #[cfg(unix)]
+    {
+        let interrupt_tx_sighup = interrupt_tx.clone();
+        tokio::spawn(async move {
+            let mut sighup = tokio::signal::unix::signal(tokio::signal::unix::SignalKind::hangup())
+                .expect("Failed to register SIGHUP handler");
+            sighup.recv().await;
+            warn!("SIGHUP received (terminal closed), terminating immediately...");
+            let _ = interrupt_tx_sighup.send(true);
+        });
+    }
+
+    // Resolve prompt content with precedence:
+    // 1. CLI -p (inline text)
+    // 2. CLI -P (file path)
+    // 3. Config prompt (inline text)
+    // 4. Config prompt_file (file path)
+    // 5. Default PROMPT.md
+    let prompt_content = resolve_prompt_content(&config.event_loop)?;
+
+    // For fresh runs (not resume), generate a unique timestamped events file
+    // This prevents stale events from previous runs polluting new runs (issue #82)
+    // The marker file `.ralph/current-events` coordinates path between Ralph and agents
+    if !resume {
+        let run_id = chrono::Utc::now().format("%Y%m%d-%H%M%S").to_string();
+        let events_path = format!(".ralph/events-{}.jsonl", run_id);
+
+        fs::create_dir_all(".ralph").context("Failed to create .ralph directory")?;
+        fs::write(".ralph/current-events", &events_path)
+            .context("Failed to write .ralph/current-events marker file")?;
+
+        debug!("Created events file for this run: {}", events_path);
+    }
+
+    // Initialize event loop
+    let mut event_loop = EventLoop::new(config.clone());
+
+    // For resume mode, we initialize with a different event topic
+    // This tells the planner to read existing scratchpad rather than creating a new one
+    if resume {
+        event_loop.initialize_resume(&prompt_content);
+    } else {
+        event_loop.initialize(&prompt_content);
+    }
+
+    // Set up session recording if requested
+    // This records all events to a JSONL file for replay testing
+    let _session_recorder: Option<Arc<SessionRecorder<BufWriter<File>>>> =
+        if let Some(record_path) = record_session {
+            let file = File::create(&record_path).with_context(|| {
+                format!("Failed to create session recording file: {:?}", record_path)
+            })?;
+            let recorder = Arc::new(SessionRecorder::new(BufWriter::new(file)));
+
+            // Record metadata for the session
+            recorder.record_meta(Record::meta_loop_start(
+                &config.event_loop.prompt_file,
+                config.event_loop.max_iterations,
+                if enable_tui { Some("tui") } else { Some("cli") },
+            ));
+
+            // Wire observer to EventBus so events are recorded
+            let observer = SessionRecorder::make_observer(Arc::clone(&recorder));
+            event_loop.add_observer(observer);
+
+            info!("Session recording enabled: {:?}", record_path);
+            Some(recorder)
+        } else {
+            None
+        };
+
+    // Initialize event logger for debugging
+    let mut event_logger = EventLogger::default_path();
+
+    // Log initial event (task.start or task.resume)
+    let (start_topic, start_triggered) = if resume {
+        ("task.resume", "planner")
+    } else {
+        ("task.start", "planner")
+    };
+    let start_event = Event::new(start_topic, &prompt_content);
+    let start_record =
+        EventRecord::new(0, "loop", &start_event, Some(&HatId::new(start_triggered)));
+    if let Err(e) = event_logger.log(&start_record) {
+        warn!("Failed to log start event: {}", e);
+    }
+
+    // Create backend from config - TUI mode uses the same backend as non-TUI
+    // The TUI is an observation layer that displays output, not a different mode
+    let backend = CliBackend::from_config(&config.cli).map_err(|e| anyhow::Error::new(e))?;
+
+    // Create PTY executor if using interactive mode
+    let mut pty_executor = if use_pty {
+        let idle_timeout_secs = if user_interactive {
+            config.cli.idle_timeout_secs
+        } else {
+            0
+        };
+        let pty_config = PtyConfig {
+            interactive: user_interactive,
+            idle_timeout_secs,
+            workspace_root: config.core.workspace_root.clone(),
+            ..PtyConfig::from_env()
+        };
+        Some(PtyExecutor::new(backend.clone(), pty_config))
+    } else {
+        None
+    };
+
+    // Create termination signal for TUI shutdown
+    let (terminated_tx, terminated_rx) = tokio::sync::watch::channel(false);
+
+    // Wire TUI with termination signal and shared state
+    // TUI is observation-only - works in both interactive and autonomous modes
+    // Only requirement: stdout must be a terminal for TUI to render
+    let enable_tui = enable_tui && stdout().is_terminal();
+    let (mut tui_handle, tui_state) = if enable_tui {
+        // Build hat map for dynamic topic-to-hat resolution
+        // This allows TUI to display custom hats (e.g., "Security Reviewer")
+        // instead of generic "ralph" for all events
+        let hat_map = build_tui_hat_map(event_loop.registry());
+        let tui = Tui::new()
+            .with_hat_map(hat_map)
+            .with_termination_signal(terminated_rx);
+
+        // Get shared state before spawning (for content streaming)
+        let state = tui.state();
+
+        // Wire interrupt channel so TUI can signal main loop on Ctrl+C
+        // (raw mode prevents SIGINT from being generated by the OS)
+        let tui = tui.with_interrupt_tx(interrupt_tx.clone());
+
+        let observer = tui.observer();
+        event_loop.add_observer(observer);
+        (
+            Some(tokio::spawn(async move { tui.run().await })),
+            Some(state),
+        )
+    } else {
+        (None, None)
+    };
+
+    // Give TUI task time to initialize (enter alternate screen, enable raw mode)
+    // before the main loop starts doing work
+    if tui_handle.is_some() {
+        tokio::time::sleep(Duration::from_millis(100)).await;
+    }
+
+    // Log execution mode - hat info already logged by initialize()
+    let exec_mode = if user_interactive {
+        "interactive"
+    } else {
+        "autonomous"
+    };
+    debug!(execution_mode = %exec_mode, "Execution mode configured");
+
+    // Track the last hat to detect hat changes for logging
+    let mut last_hat: Option<HatId> = None;
+
+    // Track consecutive fallback attempts to prevent infinite loops
+    let mut consecutive_fallbacks: u32 = 0;
+    const MAX_FALLBACK_ATTEMPTS: u32 = 3;
+
+    // Helper closure to handle termination (writes summary, prints status)
+    let handle_termination = |reason: &TerminationReason,
+                              state: &ralph_core::LoopState,
+                              scratchpad: &str| {
+        // Per spec: Write summary file on termination
+        let summary_writer = SummaryWriter::default();
+        let scratchpad_path = std::path::Path::new(scratchpad);
+        let scratchpad_opt = if scratchpad_path.exists() {
+            Some(scratchpad_path)
+        } else {
+            None
+        };
+
+        // Get final commit SHA if available
+        let final_commit = get_last_commit_info();
+
+        if let Err(e) = summary_writer.write(reason, state, scratchpad_opt, final_commit.as_deref())
+        {
+            warn!("Failed to write summary file: {}", e);
+        }
+
+        // Print termination info to console (skip in TUI mode - TUI handles display)
+        if !enable_tui {
+            print_termination(reason, state, use_colors);
+        }
+    };
+
+    // Main orchestration loop
+    loop {
+        // Check for interrupt signal at start of each iteration
+        // This catches TUI Ctrl+C (via interrupt_tx) before printing iteration separator
+        if *interrupt_rx.borrow() {
+            #[cfg(unix)]
+            {
+                use nix::sys::signal::{Signal, killpg};
+                use nix::unistd::getpgrp;
+                let pgid = getpgrp();
+                debug!(
+                    "Interrupt detected at loop start, sending SIGTERM to process group {}",
+                    pgid
+                );
+                let _ = killpg(pgid, Signal::SIGTERM);
+                tokio::time::sleep(Duration::from_millis(250)).await;
+                let _ = killpg(pgid, Signal::SIGKILL);
+            }
+            let reason = TerminationReason::Interrupted;
+            let terminate_event = event_loop.publish_terminate_event(&reason);
+            log_terminate_event(
+                &mut event_logger,
+                event_loop.state().iteration,
+                &terminate_event,
+            );
+            handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+            // Signal TUI to exit immediately on interrupt
+            let _ = terminated_tx.send(true);
+            return Ok(reason);
+        }
+
+        // Check termination before execution
+        if let Some(reason) = event_loop.check_termination() {
+            // Per spec: Publish loop.terminate event to observers
+            let terminate_event = event_loop.publish_terminate_event(&reason);
+            log_terminate_event(
+                &mut event_logger,
+                event_loop.state().iteration,
+                &terminate_event,
+            );
+            handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+            // Wait for user to exit TUI (press 'q') on natural completion
+            if let Some(handle) = tui_handle.take() {
+                let _ = handle.await;
+            }
+            return Ok(reason);
+        }
+
+        // Get next hat to execute, with fallback recovery if no pending events
+        let hat_id = match event_loop.next_hat() {
+            Some(id) => {
+                // Reset fallback counter on successful event routing
+                consecutive_fallbacks = 0;
+                id.clone()
+            }
+            None => {
+                // No pending events - try to recover by injecting a fallback event
+                // This triggers the built-in planner to assess the situation
+                consecutive_fallbacks += 1;
+
+                if consecutive_fallbacks > MAX_FALLBACK_ATTEMPTS {
+                    warn!(
+                        attempts = consecutive_fallbacks,
+                        "Fallback recovery exhausted after {} attempts, terminating",
+                        MAX_FALLBACK_ATTEMPTS
+                    );
+                    let reason = TerminationReason::Stopped;
+                    let terminate_event = event_loop.publish_terminate_event(&reason);
+                    log_terminate_event(
+                        &mut event_logger,
+                        event_loop.state().iteration,
+                        &terminate_event,
+                    );
+                    handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+                    // Wait for user to exit TUI (press 'q') on natural completion
+                    if let Some(handle) = tui_handle.take() {
+                        let _ = handle.await;
+                    }
+                    return Ok(reason);
+                }
+
+                if event_loop.inject_fallback_event() {
+                    // Fallback injected successfully, continue to next iteration
+                    // The planner will be triggered and can either:
+                    // - Dispatch more work if tasks remain
+                    // - Output LOOP_COMPLETE if done
+                    // - Determine what went wrong and recover
+                    continue;
+                }
+
+                // Fallback not possible (no planner hat or doesn't subscribe to task.resume)
+                warn!("No hats with pending events and fallback not available, terminating");
+                let reason = TerminationReason::Stopped;
+                // Per spec: Publish loop.terminate event to observers
+                let terminate_event = event_loop.publish_terminate_event(&reason);
+                log_terminate_event(
+                    &mut event_logger,
+                    event_loop.state().iteration,
+                    &terminate_event,
+                );
+                handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+                // Wait for user to exit TUI (press 'q') on natural completion
+                if let Some(handle) = tui_handle.take() {
+                    let _ = handle.await;
+                }
+                return Ok(reason);
+            }
+        };
+
+        let iteration = event_loop.state().iteration + 1;
+
+        // Determine which hat to display in iteration separator
+        // When Ralph is coordinating (hat_id == "ralph"), show the active hat being worked on
+        let display_hat = if hat_id.as_str() == "ralph" {
+            event_loop.get_active_hat_id()
+        } else {
+            hat_id.clone()
+        };
+
+        // Per spec: Print iteration demarcation separator
+        // "Each iteration must be clearly demarcated in the output so users can
+        // visually distinguish where one iteration ends and another begins."
+        // Skip when TUI is enabled - TUI has its own header showing iteration info
+        if tui_state.is_none() {
+            print_iteration_separator(
+                iteration,
+                display_hat.as_str(),
+                event_loop.state().elapsed(),
+                config.event_loop.max_iterations,
+                use_colors,
+            );
+        }
+
+        // Log hat changes with appropriate messaging
+        // Skip in TUI mode - TUI shows hat info in header, and stdout would corrupt display
+        if last_hat.as_ref() != Some(&hat_id) {
+            if tui_state.is_none() {
+                if hat_id.as_str() == "ralph" {
+                    info!("I'm Ralph. Let's do this.");
+                } else {
+                    info!("Putting on my {} hat.", hat_id);
+                }
+            }
+            last_hat = Some(hat_id.clone());
+        }
+        debug!(
+            "Iteration {}/{} - {} active",
+            iteration, config.event_loop.max_iterations, hat_id
+        );
+
+        // Build prompt for this hat
+        let prompt = match event_loop.build_prompt(&hat_id) {
+            Some(p) => p,
+            None => {
+                error!("Failed to build prompt for hat '{}'", hat_id);
+                continue;
+            }
+        };
+
+        // In verbose mode, print the full prompt before execution
+        if verbosity == Verbosity::Verbose {
+            eprintln!("\n{}", "=".repeat(80));
+            eprintln!("PROMPT FOR {} (iteration {})", hat_id, iteration);
+            eprintln!("{}", "-".repeat(80));
+            eprintln!("{}", prompt);
+            eprintln!("{}\n", "=".repeat(80));
+        }
+
+        // Execute the prompt (interactive or autonomous mode)
+        // Get per-adapter timeout from config
+        let timeout_secs = config.adapter_settings(&config.cli.backend).timeout;
+        let timeout = Some(Duration::from_secs(timeout_secs));
+
+        // For TUI mode, get the shared lines buffer for this iteration.
+        // The buffer is owned by TuiState's IterationBuffer, so writes from
+        // TuiStreamHandler appear immediately in the TUI (real-time streaming).
+        let tui_lines: Option<Arc<std::sync::Mutex<Vec<ratatui::text::Line<'static>>>>> =
+            if let Some(ref state) = tui_state {
+                // Start new iteration and get handle to the LATEST iteration's lines buffer.
+                // We must use latest_iteration_lines_handle() instead of current_iteration_lines_handle()
+                // because the user may be viewing an older iteration while a new one executes.
+                if let Ok(mut s) = state.lock() {
+                    s.start_new_iteration();
+                    s.latest_iteration_lines_handle()
+                } else {
+                    None
+                }
+            } else {
+                None
+            };
+
+        // Race execution against interrupt signal for immediate termination on Ctrl+C
+        let mut interrupt_rx_clone = interrupt_rx.clone();
+        let interrupt_rx_for_pty = interrupt_rx.clone();
+        let tui_lines_for_pty = tui_lines.clone();
+        let execute_future = async {
+            if use_pty {
+                execute_pty(
+                    pty_executor.as_mut(),
+                    &backend,
+                    &config,
+                    &prompt,
+                    user_interactive,
+                    interrupt_rx_for_pty,
+                    verbosity,
+                    tui_lines_for_pty,
+                )
+                .await
+            } else {
+                let executor = CliExecutor::new(backend.clone());
+                let result = executor
+                    .execute(&prompt, stdout(), timeout, verbosity == Verbosity::Verbose)
+                    .await?;
+                Ok(ExecutionOutcome {
+                    output: result.output,
+                    success: result.success,
+                    termination: None,
+                })
+            }
+        };
+
+        let outcome = tokio::select! {
+            result = execute_future => result?,
+            _ = interrupt_rx_clone.changed() => {
+                // Immediately terminate children via process group signal
+                #[cfg(unix)]
+                {
+                    use nix::sys::signal::{killpg, Signal};
+                    use nix::unistd::getpgrp;
+                    let pgid = getpgrp();
+                    debug!("Sending SIGTERM to process group {}", pgid);
+                    let _ = killpg(pgid, Signal::SIGTERM);
+
+                    // Wait briefly for graceful exit, then SIGKILL
+                    tokio::time::sleep(Duration::from_millis(250)).await;
+                    let _ = killpg(pgid, Signal::SIGKILL);
+                }
+
+                let reason = TerminationReason::Interrupted;
+                let terminate_event = event_loop.publish_terminate_event(&reason);
+                log_terminate_event(&mut event_logger, event_loop.state().iteration, &terminate_event);
+                handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+                // Signal TUI to exit immediately on interrupt
+                let _ = terminated_tx.send(true);
+                return Ok(reason);
+            }
+        };
+
+        if let Some(reason) = outcome.termination {
+            let terminate_event = event_loop.publish_terminate_event(&reason);
+            log_terminate_event(
+                &mut event_logger,
+                event_loop.state().iteration,
+                &terminate_event,
+            );
+            handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+            // Wait for user to exit TUI (press 'q') on natural completion
+            if let Some(handle) = tui_handle.take() {
+                let _ = handle.await;
+            }
+            return Ok(reason);
+        }
+
+        let output = outcome.output;
+        let success = outcome.success;
+
+        // Note: TUI lines are now written directly to IterationBuffer during streaming,
+        // so no post-execution transfer is needed.
+
+        // Log events from output before processing
+        log_events_from_output(
+            &mut event_logger,
+            iteration,
+            &hat_id,
+            &output,
+            event_loop.registry(),
+        );
+
+        // Process output
+        if let Some(reason) = event_loop.process_output(&hat_id, &output, success) {
+            // Per spec: Log "All done! {promise} detected." when completion promise found
+            if reason == TerminationReason::CompletionPromise {
+                info!(
+                    "All done! {} detected.",
+                    config.event_loop.completion_promise
+                );
+            }
+            // Per spec: Publish loop.terminate event to observers
+            let terminate_event = event_loop.publish_terminate_event(&reason);
+            log_terminate_event(
+                &mut event_logger,
+                event_loop.state().iteration,
+                &terminate_event,
+            );
+            handle_termination(&reason, event_loop.state(), &config.core.scratchpad);
+            // Wait for user to exit TUI (press 'q') on natural completion
+            if let Some(handle) = tui_handle.take() {
+                let _ = handle.await;
+            }
+            return Ok(reason);
+        }
+
+        // Read events from JSONL that agent may have written
+        if let Err(e) = event_loop.process_events_from_jsonl() {
+            warn!(error = %e, "Failed to read events from JSONL");
+        }
+
+        // Precheck validation: Warn if no pending events after processing output
+        // Per EventLoop doc: "Use has_pending_events after process_output to detect
+        // if the LLM failed to publish an event."
+        if !event_loop.has_pending_events() {
+            let expected = event_loop.get_hat_publishes(&hat_id);
+            debug!(
+                hat = %hat_id.as_str(),
+                expected_topics = ?expected,
+                "No pending events after iteration. Agent may have failed to publish a valid event. \
+                 Expected one of: {:?}. Loop will terminate on next iteration.",
+                expected
+            );
+        }
+
+        // Note: Interrupt handling moved into tokio::select! above for immediate termination
+    }
+}
+
+/// Executes a prompt in PTY mode with raw terminal handling.
+/// Converts PTY termination type to loop termination reason.
+///
+/// In interactive mode, idle timeout signals "iteration complete" rather than
+/// "loop stopped", allowing the event loop to process output and continue.
+///
+/// # Arguments
+/// * `termination_type` - The PTY executor's termination type
+/// * `interactive` - Whether running in interactive mode
+///
+/// # Returns
+/// * `None` - Continue processing (iteration complete)
+/// * `Some(TerminationReason)` - Stop the loop
+fn convert_termination_type(
+    termination_type: ralph_adapters::TerminationType,
+    interactive: bool,
+) -> Option<TerminationReason> {
+    match termination_type {
+        ralph_adapters::TerminationType::Natural => None,
+        ralph_adapters::TerminationType::IdleTimeout => {
+            if interactive {
+                // In interactive mode, idle timeout signals iteration complete,
+                // not loop termination. Let output be processed for events.
+                info!("PTY idle timeout in interactive mode, iteration complete");
+                None
+            } else {
+                warn!("PTY idle timeout reached, terminating loop");
+                Some(TerminationReason::Stopped)
+            }
+        }
+        ralph_adapters::TerminationType::UserInterrupt
+        | ralph_adapters::TerminationType::ForceKill => Some(TerminationReason::Interrupted),
+    }
+}
+
+async fn execute_pty(
+    executor: Option<&mut PtyExecutor>,
+    backend: &CliBackend,
+    config: &RalphConfig,
+    prompt: &str,
+    interactive: bool,
+    interrupt_rx: tokio::sync::watch::Receiver<bool>,
+    verbosity: Verbosity,
+    tui_lines: Option<Arc<std::sync::Mutex<Vec<ratatui::text::Line<'static>>>>>,
+) -> Result<ExecutionOutcome> {
+    use crossterm::terminal::{disable_raw_mode, enable_raw_mode};
+
+    // Use provided executor or create a new one
+    // If executor is provided, TUI is connected and owns raw mode management
+    let tui_connected = executor.is_some();
+    let mut temp_executor;
+    let exec = if let Some(e) = executor {
+        e
+    } else {
+        let idle_timeout_secs = if interactive {
+            config.cli.idle_timeout_secs
+        } else {
+            0
+        };
+        let pty_config = PtyConfig {
+            interactive,
+            idle_timeout_secs,
+            workspace_root: config.core.workspace_root.clone(),
+            ..PtyConfig::from_env()
+        };
+        temp_executor = PtyExecutor::new(backend.clone(), pty_config);
+        &mut temp_executor
+    };
+
+    // Set TUI mode flag when TUI is connected (tui_lines is Some)
+    // This replaces the broken output_rx.is_none() detection in PtyExecutor
+    if tui_lines.is_some() {
+        exec.set_tui_mode(true);
+    }
+
+    // Enter raw mode for interactive mode to capture keystrokes
+    // Skip if TUI is connected - TUI owns raw mode and will manage it
+    if interactive && !tui_connected {
+        enable_raw_mode().context("Failed to enable raw mode")?;
+    }
+
+    // Use scopeguard to ensure raw mode is restored on any exit path
+    // Skip if TUI is connected - TUI owns raw mode
+    let _guard = scopeguard::guard((interactive, tui_connected), |(is_interactive, tui)| {
+        if is_interactive && !tui {
+            let _ = disable_raw_mode();
+        }
+    });
+
+    // Run PTY executor with shared interrupt channel
+    let result = if interactive && tui_lines.is_none() {
+        // Raw interactive mode only when not using TUI (TUI handles its own terminal)
+        exec.run_interactive(prompt, interrupt_rx).await
+    } else if let Some(lines) = tui_lines {
+        // TUI mode: use TuiStreamHandler to capture output for TUI display
+        let verbose = verbosity == Verbosity::Verbose;
+        let mut handler = TuiStreamHandler::with_lines(verbose, lines);
+        exec.run_observe_streaming(prompt, interrupt_rx, &mut handler)
+            .await
+    } else {
+        // Use streaming handler for non-interactive mode (respects verbosity)
+        // Use PrettyStreamHandler for StreamJson backends (Claude) on TTY for markdown rendering
+        // Use ConsoleStreamHandler for Text format backends (Kiro, Gemini, etc.) for immediate output
+        let use_pretty =
+            backend.output_format == BackendOutputFormat::StreamJson && stdout().is_terminal();
+
+        match verbosity {
+            Verbosity::Quiet => {
+                let mut handler = QuietStreamHandler;
+                exec.run_observe_streaming(prompt, interrupt_rx, &mut handler)
+                    .await
+            }
+            Verbosity::Normal => {
+                if use_pretty {
+                    let mut handler = PrettyStreamHandler::new(false);
+                    exec.run_observe_streaming(prompt, interrupt_rx, &mut handler)
+                        .await
+                } else {
+                    let mut handler = ConsoleStreamHandler::new(false);
+                    exec.run_observe_streaming(prompt, interrupt_rx, &mut handler)
+                        .await
+                }
+            }
+            Verbosity::Verbose => {
+                if use_pretty {
+                    let mut handler = PrettyStreamHandler::new(true);
+                    exec.run_observe_streaming(prompt, interrupt_rx, &mut handler)
+                        .await
+                } else {
+                    let mut handler = ConsoleStreamHandler::new(true);
+                    exec.run_observe_streaming(prompt, interrupt_rx, &mut handler)
+                        .await
+                }
+            }
+        }
+    };
+
+    match result {
+        Ok(pty_result) => {
+            let termination = convert_termination_type(pty_result.termination, interactive);
+
+            // Use extracted_text for event parsing when available (NDJSON backends like Claude),
+            // otherwise fall back to stripped_output (non-JSON backends or interactive mode).
+            // This fixes event parsing for Claude's stream-json output where event tags like
+            // <event topic="..."> are inside JSON string values and not directly visible.
+            let output_for_parsing = if pty_result.extracted_text.is_empty() {
+                pty_result.stripped_output
+            } else {
+                pty_result.extracted_text
+            };
+            Ok(ExecutionOutcome {
+                output: output_for_parsing,
+                success: pty_result.success,
+                termination,
+            })
+        }
+        Err(e) => {
+            // PTY allocation may have failed - log and continue with error
+            warn!("PTY execution failed: {}, continuing with error status", e);
+            Err(anyhow::Error::new(e))
+        }
+    }
+}
+
+/// Logs events parsed from output to the event history file.
+fn log_events_from_output(
+    logger: &mut EventLogger,
+    iteration: u32,
+    hat_id: &HatId,
+    output: &str,
+    registry: &ralph_core::HatRegistry,
+) {
+    let parser = EventParser::new();
+    let events = parser.parse(output);
+
+    for event in events {
+        // Determine which hat will be triggered by this event
+        let triggered = registry.find_by_trigger(event.topic.as_str());
+
+        // Per spec: Log "Published {topic} -> triggers {hat}" at DEBUG level
+        if let Some(triggered_hat) = triggered {
+            debug!("Published {} -> triggers {}", event.topic, triggered_hat);
+        } else {
+            debug!("Published {} -> no hat triggered", event.topic);
+        }
+
+        let record = EventRecord::new(iteration, hat_id.to_string(), &event, triggered);
+
+        if let Err(e) = logger.log(&record) {
+            warn!("Failed to log event {}: {}", event.topic, e);
+        }
+    }
+}
+
+/// Logs the loop.terminate system event to the event history.
+///
+/// Per spec: loop.terminate is an observer-only event published on loop exit.
+fn log_terminate_event(logger: &mut EventLogger, iteration: u32, event: &Event) {
+    // loop.terminate is published by the orchestrator, not a hat
+    // No hat can trigger on it (it's observer-only)
+    let record = EventRecord::new(iteration, "loop", event, None::<&HatId>);
+
+    if let Err(e) = logger.log(&record) {
+        warn!("Failed to log loop.terminate event: {}", e);
+    }
+}
+
+/// Gets the last commit info (short SHA and subject) for the summary file.
+fn get_last_commit_info() -> Option<String> {
+    let output = Command::new("git")
+        .args(["log", "-1", "--format=%h: %s"])
+        .output()
+        .ok()?;
+
+    if output.status.success() {
+        let info = String::from_utf8_lossy(&output.stdout).trim().to_string();
+        if info.is_empty() { None } else { Some(info) }
+    } else {
+        None
+    }
+}
+
+/// Resolves prompt content with proper precedence.
+///
+/// Precedence (highest to lowest):
+/// 1. CLI -p "text" (inline prompt text)
+/// 2. CLI -P path (prompt file path)
+/// 3. Config event_loop.prompt (inline prompt text)
+/// 4. Config event_loop.prompt_file (prompt file path)
+/// 5. Default PROMPT.md
+///
+/// Note: CLI overrides are already applied to config before this function is called.
+fn resolve_prompt_content(event_loop_config: &ralph_core::EventLoopConfig) -> Result<String> {
+    debug!(
+        inline_prompt = ?event_loop_config.prompt.as_ref().map(|s| format!("{}...", &s[..s.len().min(50)])),
+        prompt_file = %event_loop_config.prompt_file,
+        "Resolving prompt content"
+    );
+
+    // Check for inline prompt first (CLI -p or config prompt)
+    if let Some(ref inline_text) = event_loop_config.prompt {
+        debug!(len = inline_text.len(), "Using inline prompt text");
+        return Ok(inline_text.clone());
+    }
+
+    // Check for prompt file (CLI -P or config prompt_file or default)
+    let prompt_file = &event_loop_config.prompt_file;
+    if !prompt_file.is_empty() {
+        let path = std::path::Path::new(prompt_file);
+        debug!(path = %prompt_file, exists = path.exists(), "Checking prompt file");
+        if path.exists() {
+            let content = std::fs::read_to_string(path)
+                .with_context(|| format!("Failed to read prompt file: {}", prompt_file))?;
+            debug!(path = %prompt_file, len = content.len(), "Read prompt from file");
+            return Ok(content);
+        } else {
+            // File specified but doesn't exist - error with helpful message
+            anyhow::bail!(
+                "Prompt file '{}' not found. Check the path or use -p \"text\" for inline prompt.",
+                prompt_file
+            );
+        }
+    }
+
+    // No valid prompt source found
+    anyhow::bail!(
+        "No prompt specified. Use -p \"text\" for inline prompt, -P path for file, \
+         or create PROMPT.md in the current directory."
+    )
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_pty_always_enabled_for_streaming() {
+        // PTY mode is always enabled for real-time streaming output.
+        // This ensures all backends (claude, gemini, kiro, codex, amp) get
+        // streaming output instead of buffered output from CliExecutor.
+        let use_pty = true; // Matches the actual implementation
+
+        // PTY should always be true regardless of backend or mode
+        assert!(use_pty, "PTY should always be enabled for streaming output");
+    }
+
+    #[test]
+    fn test_user_interactive_mode_determination() {
+        // user_interactive is determined by default_mode setting, not PTY.
+        // PTY handles output streaming; user_interactive handles input forwarding.
+
+        // Autonomous mode: no user input forwarding
+        let autonomous_interactive = false;
+        assert!(
+            !autonomous_interactive,
+            "Autonomous mode should not forward user input"
+        );
+
+        // Interactive mode with TTY: forward user input
+        let interactive_with_tty = true;
+        assert!(
+            interactive_with_tty,
+            "Interactive mode with TTY should forward user input"
+        );
+    }
+
+    #[test]
+    fn test_idle_timeout_interactive_mode_continues() {
+        // Given: interactive mode and IdleTimeout termination
+        let termination_type = ralph_adapters::TerminationType::IdleTimeout;
+        let interactive = true;
+
+        // When: converting termination type
+        let result = convert_termination_type(termination_type, interactive);
+
+        // Then: should return None (allow iteration to continue)
+        assert!(
+            result.is_none(),
+            "Interactive mode idle timeout should return None to allow iteration progression"
+        );
+    }
+
+    #[test]
+    fn test_idle_timeout_autonomous_mode_stops() {
+        // Given: autonomous mode and IdleTimeout termination
+        let termination_type = ralph_adapters::TerminationType::IdleTimeout;
+        let interactive = false;
+
+        // When: converting termination type
+        let result = convert_termination_type(termination_type, interactive);
+
+        // Then: should return Some(Stopped)
+        assert_eq!(
+            result,
+            Some(TerminationReason::Stopped),
+            "Autonomous mode idle timeout should return Stopped"
+        );
+    }
+
+    #[test]
+    fn test_natural_termination_always_continues() {
+        // Given: Natural termination in any mode
+        let termination_type = ralph_adapters::TerminationType::Natural;
+
+        // When/Then: should return None regardless of mode
+        assert!(
+            convert_termination_type(termination_type.clone(), true).is_none(),
+            "Natural termination should continue in interactive mode"
+        );
+        assert!(
+            convert_termination_type(termination_type, false).is_none(),
+            "Natural termination should continue in autonomous mode"
+        );
+    }
+
+    #[test]
+    fn test_user_interrupt_always_terminates() {
+        // Given: UserInterrupt termination in any mode
+        let termination_type = ralph_adapters::TerminationType::UserInterrupt;
+
+        // When/Then: should return Interrupted regardless of mode
+        assert_eq!(
+            convert_termination_type(termination_type.clone(), true),
+            Some(TerminationReason::Interrupted),
+            "UserInterrupt should terminate in interactive mode"
+        );
+        assert_eq!(
+            convert_termination_type(termination_type, false),
+            Some(TerminationReason::Interrupted),
+            "UserInterrupt should terminate in autonomous mode"
+        );
+    }
+
+    #[test]
+    fn test_force_kill_always_terminates() {
+        // Given: ForceKill termination in any mode
+        let termination_type = ralph_adapters::TerminationType::ForceKill;
+
+        // When/Then: should return Interrupted regardless of mode
+        assert_eq!(
+            convert_termination_type(termination_type.clone(), true),
+            Some(TerminationReason::Interrupted),
+            "ForceKill should terminate in interactive mode"
+        );
+        assert_eq!(
+            convert_termination_type(termination_type, false),
+            Some(TerminationReason::Interrupted),
+            "ForceKill should terminate in autonomous mode"
+        );
+    }
+}
